+++ apps/bsadmin/src/app/(pages)/(set)/es-pr-modify/AnlsTypeMcVal.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { useFormContext } from "react-hook-form";
+
+const AnlsTypeMcVal = () => {
+  const { getValues } = useFormContext();
+  return <div>{getValues("anlsTypeMcVal")}</div>;
+};
+
+export default AnlsTypeMcVal;
